pyproject.toml
[tool.poetry]
name = "face_auth_api"
version = "0.1.0"
description = ""
authors = ["Roman Prudnikov <romanprudnikov1990@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
python-multipart = "^0.0.6"
onnxruntime = "^1.16"     # или onnxruntime-gpu, если GPU есть
Pillow = "^10.0"
pymilvus = "^2.5.14"
python-dotenv = "^1.0"
loguru = "^0.7"
Jinja2 = "^3.1.2"
insightface = "^0.7.3"
numpy = "^2.3.2"
pydantic = ">=2.11.0,<3.0"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
black = "^22.8.0"
mypy = "^0.982"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.19.0"
pytest-mock = "^3.11.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
